Update article.md
That English is not idiomatic at all.  You could says:

"Why don't you take a math class?"
or
"How about taking a math class?"

Also, you may want to consider using American English, and in this case, please use "math" not "maths".  The US is a bigger market than the UK, is more widely used, the US is the center of the tech industry, etc.  The standard for using English-language variables in code comes from the US, not England.  And British-English is annoying. ;) :)