Update article.md
Actually removing articles ("the"s). Rare! Usage was inconsistent from line to line -- some had "the node" and some just had "node". I'd say in this case, since `node` is formatted as code, it's intended to be an identifier (rather than a type of thing), so should stand alone. That is, if `node` **stands** for the node, don't use "the", but if 'node' just means "some node", do use "the node".