Update article.md
Small punctuation change. 

Separately, I didn't change anything on this today, but note that line 34 above:
"So to summarize: the executor runs automatically and attempts to perform a job. When it is finished with the attempt it calls `resolve` if it was successful or `reject` if there was an error."
was, before it was changed recently, pretty much the same as line 82 above:
"To summarize, the executor should perform a job (usually something that takes time) and then call `resolve` or `reject` to change the state of the corresponding promise object."
So maybe some revision there is in order -- to not have both, or to change one or both. Not familiar enough with the article at this point to make the revision.