Change 'brackets' to 'parentheses'
I think this is the way it is typically used here and elsewhere. reserving "brackets" without a qualifier for indicating square brackets., so probably no explanation is necessary. 

But, for those who want to know more:

It does seem to be the case that in British English "brackets" typically means "round brackets", which are parentheses, but in American English "brackets" typically means square brackets ("[]"). 

Admittedly, "brackets" is easier to spell. And to pronounce. Even to type, once you get used to it.