Update article.md
Hi guys, I looked at your article to solve my CORS issue but it didn't work. So eventually I found a question on StackOverflow that explained in order to make things work properly, I needed to add also "Access-Control-Allow-Origin: https://javascript.info" to the preflight response. Did anyone have the same issue? If so, we need to update the article in order to save the time of future readers :)

StackOverflow question: https://stackoverflow.com/questions/42558221/how-to-cors-enable-apache-web-server-including-preflight-and-custom-headers