Change `var` to `let` in 7.16 regexp-sticky
Commit b493f160e added a hint about which word would be found by
a certain regexp:

"A call to str.match(/\w+/) will find only the first word in the line
(var). That’s not it."

Unfortunately the result is incorrect: the regexp would find `let`, not
`var`.

Updating the hint should resolve any confusion. The example has been
tested in the console to ensure that the regexp does indeed return
`let`.